Implement a meaningful Go port of libcap
This version of the Go package libcap/cap works well
enough to be used by others. Unfortunately, to use it
we need to apply something like the included patch
(contrib/go.patch) to the build sources for the Go
runtime and syscall packages.

I'll be trying to get these accepted by the Go team
in parallel.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>