Realign around the evolving Go support for POSIX semantics syscalls.
I've moved my go.patch to address:

  https://github.com/golang/go/issues/1435

into a development patch against the upstream Go sources:

  https://go-review.googlesource.com/c/go/+/210639/

and the review process will likely evolve it somewhat. I plan to
ensure that working libcap/cap Go package is in sync with the
working state of the above development change.

As such, there is no need to keep the patch here any more.
I'll keep the tests for now, as it isn't clear to me how the Go
source tree supports tests that require privilege yet.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>