Restructure the make files into build vs. test
Also install the Go packages if built.

Remove a default behavior of installing an inheritable bit on setcap.
I'm getting alarmed that some distributions are setting the inheritable
set to full for all users. So, I don't want to provide a vector for
a trivial exploit, and hope they are not reinventing this:

https://sites.google.com/site/fullycapable/Home/thesendmailcapabilitiesissue

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>