Found and fixed a deadlock in libcap/psx's use of libpsx.
It turns out Go blocks signals from Go before creating
a new thread with a call to C.pthread_create(). This
caused about 1 in 50 runs of the libcap/psx test to deadlock.
The psx code was waiting for a signal to be received by
a thread that was blocked from hearing.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>