Avoid sys/capability.h on build architecture
libcap/_makenames.c generates a build-time helper, and as such is
compiled for
the build architecture. The use of sys/capability.h which depends on an
arch-
specific Linux header eventually leads to a FTCBFS.

However, the only reason to use the header is to estimate the size of an
array
"pointers" in _makenames.c. Rather than using this guess, the array can
simply
be allocated dynamically.

Signed-off-by: Christian Kastner <ckk@kvr.at>
Signed-off-by: Andrew G. Morgan <morgan@kernel.org>