Revamp the way we package cap and psx packages.
It has been requested that we make the "libcap/cap" package into
a module that plays better with the golang ecosystem. I was holding
off until there was a golang version that contained the
runtime.AllThreadsSyscall() support, but that appears to not have
made it to 1.15, so I'm using a development build tag dependency
in otherwise static sources for the "libcap/cap" package.

My intention is that the canonical import paths for these packages
will be:

   "git.kernel.org/libs/libcap/cap"
   "git.kernel.org/libs/libcap/psx"

That being said, I may have to move them if I can't get some
proxy to resolve these paths to the right git repo of kernel.org.

The is work in the direction of addressing:

   https://bugzilla.kernel.org/show_bug.cgi?id=207567

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>