Refactored the psx package to build as a Go module.
Cleaned up the Go module redirection html file, now installed
at:

  https://kernel.org/pub/linux/libs/security/libcap/

Note, I've moved the C source for libpsx.a into the psx/
directory, but the libpsx.a file is still built in the libcap
subdirectory as before. I also symlinked the C include files from
the psx/ directory. This made the source compile in conjuction
with the "psx" Go package automatically. It also substantially
simplified the go/Makefile.

I feel pretty good about this next version from the perspective
of a viable "psx" build. Caveat the need for CGO_LDFLAGS_ALLOW
on the command line pre-go1.15. Hopefully, the psx package comment
is enough for folk to figure that detail out.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>