Trying to stop the DOS attack on me from lawyers.
Reformatted the license files to capture the same info but
in a more clear way.

Note, I have migrated the license texts to "License" files
since, apparently, LICENSE files are picked up and parsed by
all sorts of automation that I don't want to reverse engineer.

The compiled binaries refer to said LICENSE file(s).

The default LICENSE files refer to the License files for
details. It is expected that folk that build against libcap
and friends, and need to choose one or the other license will
simplify their copies of LICENSE consistent with their own
needs and the scope explained in the License file(s).

All code distributed with the official libcap, unless
explicitly stated, is covered by the libcap License file.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>