Implement cap.FuncLauncher()
Create a new launcher that does not require a specific target executable
but comes as close as Go lets to running code in a disposable security
context.

In C, this same functionality is implemented as a full blown fork() and
is useful when libcap is linked with libpsx. In Go, we can't
meaningfully fork the go runtime, so this is the next best thing. It is
intended to implement the feature discussed here:

  https://bugzilla.kernel.org/show_bug.cgi?id=211919

Previously, we banned use of capability etc setting cap.Fn() calls from
within a launch callback. With this change, we now permit the callback
to perform cap.Fn() calls in such a way that they are automatically
limited to the callback goroutine and do not affect the rest of the
process. Because this goroutine will die after the launch callback
completes this behavior does not confuse the rest of the process - which
retains a process wide security state - POSIX semantics.

The main process goroutines outside of launchers will continue to block
on cap.Fn() calls that modify process capability state until all
outstanding launch()es have completed.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>