Import an old version of su from SimplePAMApps-0.60
When Linux-PAM was getting its act together (more than two decades ago)
we cobbled together a set of system apps and made them use Linux-PAM
for authentication. Once Redhat shipped Linux-PAM, the mainstream
versions of these apps adopted Linux-PAM and these simple ones withered.
I want to explore some pam_cap.so related issues and so I've resurrected
one of them, su, which announces itself to libpam with the name "sucap".

I'm not sure where I'll go with this yet, but my first goal is to
reproduce the issue:

   https://bugzilla.kernel.org/show_bug.cgi?id=212945

and validate the workaround I've added to that module is sufficient.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>