Add an autoauth module flag to pam_cap.so
If present, this causes pam_cap.so to return PAM_SUCCESS if
the module is invoked as pam_sm_authenticate() and a PAM_USER
is known. Otherwise, the return status is dependent on whether
or not the PAM_USER is covered by a configured rule.

This feature addresses this feature request:

  https://bugzilla.kernel.org/show_bug.cgi?id=213279

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>