Add the captree example.
This is a small command line utility for doing something like pstree
but focused on revealing the full capability state of the processes
and threads shown.

This requires support provided in the cap.IABGetPID() function which
will debut in libcap-2.54. For now, the binary is only buildable from
HEAD in the git repository.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>