Stop installing the cap and psx packages, but do install captree.
We also add the cap.ProcRoot() API to let the user redirect to their
local /proc/ directory - in case anyone runs with an unusual setup
like that.

I've been studying the downstream package definitions and no one
it doesn't seem popular to build the Go packages. Indeed, Go folk
themselves prefer to install via modules anyway, so we're getting
with the program.

However, if folk want to build test the Go stuff as part of a package
build and run an install as well, we reward them with the 'captree'
binary.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>