Add some debugging info for the pam_cap.so deferred callback.
As with the other D(()) entries in the pam_cap.so module, this
is enabled if the /* #define PAM_DEBUG */ comment is uncommented
at the top of the pam_cap.so file.

I tried this on a sample app and it didn't actually follow the
documentation:

http://www.linux-pam.org/Linux-PAM-html/adg-interface-by-app-expected.html#adg-pam_end

where no pam_end() call was made to terminate the fork()ed copy of the pamh
value. That app needs to be fixed.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>