An example of a shared library object with its own file capability.
I've been exploring the idea of how to create limited use privileged
binaries that can be linked into otherwise unprivileged binaries.
This is a worked example of the bootstrapping process for a webserver.

I intend to provide a more complete writeup of what is going on with this
example here:

  https://sites.google.com/site/fullycapable/capable-shared-objects

For this present example to work you have to be using a libcap that
includes cap_launch support (ie., libcap 2.33+, but this code will be
included with libcap-2.61 and might inadvertently actually require
something that new to work robustly).

This code appears to be very fragile at present. It works on my
Chromebook's linux container, but not under Fedora 34 - segfaulting.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>