Work around a __i386__ compilation issue for runnable .so files.
This was reported by Sam James and debugged with respect to:

  https://bugs.gentoo.org/show_bug.cgi?id=820071

Modern versions of glibc employ SSE instructions that require the
stack to be aligned to 16 bytes in order to execute movaps and
friends to stack stored memory. The ABI for x86_64 requires this
alignment so we'd not seen this issue before being cc:d into the
bug.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>