Modify DYNAMIC=no linking to only refer to libcap.a
There is a longstanding WONT_FIX bug:

https://sourceware.org/bugzilla/show_bug.cgi?id=12491

that has been causing capsh, when linked fully statically,
to segfault. So, for non-dynamic linking of capsh etc utilities
only link statically to libcap. This way, in tree builds can be
guaranteed to get to execute with in tree API changes. For
normal installations, DYNAMIC=yes works as before.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>