Add an example of combining Go, C code and "psx" without cgo.
This example was developed while investigating the issues discussed in:

  https://bugzilla.kernel.org/show_bug.cgi?id=216610

At this time, it is not possible to build CGO_ENABLED=1 and include
the "psx" package without using its "cgo"-tagged build variant.

This example provides a worked example of doing the opposite: link a
CGO_ENABLED=0 binary with "psx", including some compiled C code.

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>