Add support for optimized C compilation to .syso objects.
It took me a while to figure out why optimized C compilation seemed
to generate miscomputation of the Fibonacci number sequence. It appears
to be an unresolved issue with Go's internal linking which is discussed
here:

  https://github.com/golang/go/issues/24321

For a compute kernel, it seems important to be able to accommodate
compiler optimization. This adds some refinement for the strategy
I'm exploring to address:

  https://bugzilla.kernel.org/show_bug.cgi?id=216610

Signed-off-by: Andrew G. Morgan <morgan@kernel.org>